requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.3.0
pypng==0.20220715.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.11.0
uritemplate==4.1.1
